Remove old, untested client
