+++ internal/bluesnews/client_test.go
